split HorizontalBar components into 2 new classes
